chore: move demo folder out of src
