+++ demo/data/logo.js
