+++ demo/examples/GeoChart.vue
@@ -8,7 +8,7 @@ import {
   TooltipComponent
 } from "echarts/components";
 import { shallowRef } from "vue";
-import VChart from "../../ECharts";
+import VChart from "../../src/ECharts";
 import VExample from "./Example.vue";
 import getData from "../data/map";
 import chinaMap from "../china.json";
