+++ demo/examples/PieChart.vue
@@ -8,7 +8,7 @@ import {
   TooltipComponent
 } from "echarts/components";
 import { shallowRef, onMounted } from "vue";
-import VChart from "../../ECharts";
+import VChart from "../../src/ECharts";
 import VExample from "./Example.vue";
 import getData from "../data/pie";
