+++ demo/examples/RadarChart.vue
@@ -7,7 +7,7 @@ import {
   TooltipComponent
 } from "echarts/components";
 import { shallowRef } from "vue";
-import VChart from "../../ECharts";
+import VChart from "../../src/ECharts";
 import VExample from "./Example.vue";
 import { useScoreStore } from "../data/radar";
