+++ demo/examples/ScatterChart.vue
@@ -7,7 +7,7 @@ import {
   LegendComponent
 } from "echarts/components";
 import { shallowRef } from "vue";
-import VChart from "../../ECharts";
+import VChart from "../../src/ECharts";
 import VExample from "./Example.vue";
 import getData from "../data/scatter";
